Task 2/4.2
Updated database settings for testing. Errors with connecting to the database in tests have been fixed. Added endpoint for obtaining statistics about users. Added methods to UserService to count users registered in the last 7 days, to get the top 5 users with the longest names, and to get the percentage of users registered in a particular domain